ContractDraftingWebApp/requirements.txt
django==5.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-cleanup
psycopg2-binary
docxtpl
num2words
django-filter
python-decouple
